move SourceMap.js to src; inline btoa
