+++ src/index.js
@@ -2,4 +2,4 @@ import MagicString from './MagicString.js';
 
 export default MagicString;
 export { default as Bundle } from './Bundle.js';
-export { default as SourceMap } from './utils/SourceMap.js';
+export { default as SourceMap } from './SourceMap.js';
